Store the full domain and range for the crop values calculated from the data to make the color calculations more accurate
Prevents a bug that resulted in the gradients changing with the
threshold even when the threshold was well below the minimum value for
the regression. Also only show a band for the highest energy capture
ratio; do not overlap. Calculate the latitude domains per crop based on
the raw data.